refactor: comments and feat: my posts

+++ DigiBank_API/digibank_back/Domains/Emprestimo.cs
@@ -8,7 +8,7 @@ namespace digibank_back.Domains
     public partial class Emprestimo
     {
         public short IdEmprestimo { get; set; }
-        public short IdUsuario { get; set; }
+        public int IdUsuario { get; set; }
         public byte IdCondicao { get; set; }
         public byte IdEmprestimoOptions { get; set; }
         public decimal? ValorPago { get; set; }
